% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example Dataset}
\format{
A data frame with 263 rows (Entrez gene IDs) and 30 columns (samples)
}
\source{
Randomly sampled genes from an example transcriptomics dose-response dataset.
}
\usage{
data(example_data)
}
\description{
This is an example dataset included in the \code{fastbmdR} package. It can be used for demonstrations and testing.
}
\keyword{datasets}
