% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastbmdR_main.R
\name{FilterDRFit}
\alias{FilterDRFit}
\title{FilterDRFit}
\usage{
FilterDRFit(fitObj, lof.pval = 0.1, filt.var = "AIC.model")
}
\arguments{
\item{fitObj}{A \code{fitObj} object created by PerformCurveFitting.}

\item{lof.pval}{A \code{numeric} object that specifies the lack-of-fit p-value threshold.}

\item{filt.var}{A \code{character} object that specifies the variable used for selecting the best curve fit.}
}
\value{
An updated \code{fitObj} object with filtered curve fits.
}
\description{
This function filters curves according to specified qaqc thresholds and
selects the best-fit curve from the remaining curves for each feature
according to the specified filt.var variable. It returns an updated fitObj
object that includes the filtered curve fits.
}
