% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastbmdR_main.R
\name{PerformBMDCalc}
\alias{PerformBMDCalc}
\title{PerformBMDCalc}
\usage{
PerformBMDCalc(
  fitObj,
  ncpus = 1,
  num.sds = 1,
  bmr.method = "sample.mean",
  log10.dose = FALSE
)
}
\arguments{
\item{fitObj}{A \code{fitObj} object from FilterDRFit.}

\item{ncpus}{An \code{int} specifying the number of cpus to use in parallel.}

\item{num.sds}{A \code{numeric} specifying the number of std to define the BMR.}

\item{bmr.method}{A \code{character} specifying the BMR definition method.}

\item{log10.dose}{A \code{logical} specifying whether the dose is on a log scale.}
}
\value{
A \code{fitObjFiltBMD} with BMD results.
}
\description{
This function computes the benchmark dose (BMD) for each curve fit. The BMD
is defined as where the fitted curve reaches the benchmark response (BMR).
The BMR is defined relative to the controls values, for example as the mean
of the control values, plus or minus 1 standard deviation of the controls.
Various BMD QA/QC metrics are computed to assess BMD uncertainty. This
function returns a dataframe with the BMD fit results.
}
