% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastbmdR_main.R
\name{PerformCurveFitting}
\alias{PerformCurveFitting}
\title{PerformCurveFitting}
\usage{
PerformCurveFitting(
  data,
  dose,
  ncpus = 1,
  models = c("Exp2", "Exp3", "Exp4", "Exp5", "Lin", "Poly2", "Hill", "Power")
)
}
\arguments{
\item{data}{A \code{data.frame} with omics features in rows and sample in columns.}

\item{dose}{A \code{c()} containing the dose associated with each sample in data.}

\item{ncpus}{An \code{int} specifying the number of cpus to run in parallel.}

\item{models}{A \code{c()} of model names to use for curve fitting.}
}
\value{
A \code{fitObj} object containing the input data and curve fitting results.
}
\description{
This function fits the selected models for every feature in the input 'data'
dataframe and computes statistics describing the curve fit. It can be
executed in parallel by specifying 'ncpus'. It returns 'fitObj', a list that
includes the input data and the curve fit results.
}
