% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastbmdR_plots.R
\name{plot_bmd_curve}
\alias{plot_bmd_curve}
\title{plot_bmd_curve}
\usage{
plot_bmd_curve(feature, fitObj, dose_spacing = "natural", return_type = "plot")
}
\arguments{
\item{feature}{A \code{character} specifying which feature to plot a curve for.}

\item{fitObj}{A \code{fitObj} containing the curve fits and data.}

\item{dose_spacing}{A \code{character} specifying the scale of the dose range.
Possible values are 'natural', 'log10', and 'log2'.}

\item{return_type}{A \code{character} specifying the return data format.
Possible values are 'plot', 'plot.object', and 'plot.data'.}
}
\value{
Either a plot, ggplot grob, or dataframe.
}
\description{
This function plots both the observed data points and the fitted curve for
one feature in the fitres.filt dataframe in a fitObj.
}
