% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastbmdR_main.R
\name{scoresPOD}
\alias{scoresPOD}
\title{scoresPOD}
\usage{
scoresPOD(dat, dose, log10.dose = FALSE, num.sds = 1, filt.var = "AIC.model")
}
\arguments{
\item{dat}{A \code{data.frame} with omics features in rows and sample in columns.}

\item{dose}{A \code{c()} containing the dose associated with each sample in data.}

\item{log10.dose}{A \code{logical} specifying whether the dose is on a log scale.}

\item{num.sds}{A \code{numeric} specifying the number of standard deviations to define the BMR.}

\item{filt.var}{A \code{character} that specifies the variable used for selecting the best curve fit.}
}
\value{
A \code{data.frame} with curve fit and BMD results.
}
\description{
This function around the FastBMD workflow with parameters set to work well
for omics datasets that have been summarized as Mahalanobis distances. In
this scenario, it is important to be extra conservative with filtering curves
out because there is only a small number of features (one for global, ~15 for
categorical) thus losing marginal fits can greatly impact the results. Also,
a symmetric BMD (+/- SD) is not appropriate because distance values cannot go
below zero.
}
