% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastbmdR_main.R
\name{PerformCurveFitting}
\alias{PerformCurveFitting}
\title{Structure of "PerformCurveFitting" based on source code of DROmics.
Specific model fits (other than Hill model);
BMD calculation function; and modifications
made throughout to be consistent with NTP approach for
dose-response modeling, but
consider citing: https://pubs.acs.org/doi/10.1021/acs.est.8b04752
Some methods from here: https://doi.org/10.1371/journal.pone.0146021
PerformCurveFitting}
\usage{
PerformCurveFitting(
  data,
  dose,
  ncpus = 1,
  models = c("Exp2", "Exp3", "Exp4", "Exp5", "Lin", "Poly2", "Hill", "Power")
)
}
\arguments{
\item{data}{A \code{data.frame} with omics features in rows and sample in columns.}

\item{dose}{A \code{c()} containing the dose associated with each sample in data.}

\item{ncpus}{An \code{int} specifying the number of cpus to run in parallel.}

\item{models}{A \code{c()} of model names to use for curve fitting.}
}
\value{
A \code{fitObj} object containing input data and curve fit results.
}
\description{
This function fits the selected models for every feature in the input 'data'
dataframe and computes statistics describing the curve fit. It can be
executed in parallel by specifying 'ncpus'. It returns 'fitObj', a list that
includes the input data and the curve fit results.
}
